Added CheckerBuilder.WithDefaultRequestHeader(s)
You can now add default headers to be sent with each request.

The idea came to my head because I was working on a tool for Instagram, and EVERY request had to have a User-Agent header.
I didn't want to add it manually to all of my requests, so I thought about the HttpClient.DefaultRequestHeaders property and decided to implement it into Milky.