Added an "attempts" int parameter to the check process
So you can keep track of which attempt your combo is currently going through, and take different actions depending on that. That can be useful for websites with silent bans, like Mojang/Minecraft, where you sometimes have to try to authenticate multiple times with different IPs for the auth request to finally get "accepted".

You can also decide whether you want to increment it or not. For example, I wouldn't want to increment it on a proxy issue, as this wouldn't have been a real attempt.